changes from old branch

+++ src/app/recipes/recipes.service.ts
@@ -30,8 +30,6 @@ export class RecipesService {
               favorite: false
             }
           ))),
-          // fake lag
-          delay(2000)
         )
     .subscribe(r => r ? this.recipes.set(r) : []);
   }
@@ -46,14 +44,27 @@ export class RecipesService {
 
   addRecipe(recipe: CreateRecipe) {
     const updatedRecipes = this.HttpClient.post<Recipe[]>(this.baseUrl, recipe);
-    updatedRecipes.subscribe((updatedRecipeList) => this.recipes.set(updatedRecipeList))
+    updatedRecipes.subscribe((updatedRecipeList) => {
+      console.log('recipe list returned from backend:', updatedRecipeList);
+      return this.recipes.set(updatedRecipeList)})
   }
 
   deleteRecipe(id: string) {
-    this.HttpClient.delete<string>(`${this.baseUrl}/:id?id=${id}`).subscribe((returnMsg) => {
-      if (returnMsg === `Recipe with id ${id} was deleted.`) {
-        this.recipes.set(this.recipes().filter(r => r.id !== id));
-      }
-    })
+    if (confirm("Are you sure you want to delete this recipe?")) {
+      this.HttpClient.delete<string>(`${this.baseUrl}/:id?id=${id}`).subscribe((returnMsg) => {
+        if (returnMsg === `Recipe with id ${id} was deleted.`) {
+          this.recipes.set(this.recipes().filter(r => r.id !== id));
+          console.log(returnMsg);
+        }
+        else {
+          console.error('Failed to delete recipe:', returnMsg);
+        }
+      })
+    }
+  }
+
+  editRecipe(id: string, updatedRecipe: CreateRecipe) {
+    const updatedRecipes = this.HttpClient.put<Recipe[]>(`${this.baseUrl}/:id?id=${id}`, updatedRecipe);
+    updatedRecipes.subscribe((updatedRecipeList) => this.recipes.set(updatedRecipeList))
   }
 }
